+++ dlz.framework/src/main/java/com/dlz/framework/redis/excutor/IJedisSetExecutor.java
@@ -42,7 +42,7 @@ public interface IJedisSetExecutor extends IJedisExecutor {
             String key1 = getRedisKey(key);
             Long sadd = j.sadd(key1, values);
             if (seconds > 0) {
-                j.expire(key1, seconds);
+                j.expire(key1, (long)seconds);
             }
             return sadd;
         });
